Terminado codificacion ServicioAcademico
Implementado el metodo de recepcion de datos JsonPack® para metodos que
requieren varios id's para operar

+++ server/EscuelaSantaLucia/src/ar/com/santalucia/rest/JsonPack.java
@@ -3,18 +3,18 @@ package ar.com.santalucia.rest;
 import java.util.Vector;
 
 public class JsonPack {
-	private Vector<Integer> values;
+	private Vector<Long> values;
 	
 	public JsonPack(){
 		// Vector inicial de 3 elementos
-		values = new Vector<Integer>(3);
+		values = new Vector<Long>(3);
 	}
 
-	public Vector<Integer> getValues() {
+	public Vector<Long> getValues() {
 		return values;
 	}
 
-	public void setValues(Vector<Integer> values) {
+	public void setValues(Vector<Long> values) {
 		this.values = values;
 	}
 }
